Refactorizacion

+++ WebCoorporativaAPI/Controllers/PerfilController.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using WebCoorporativaAPI.Constant;
+using WebCoorporativaAPI.Helpers;
 using WebCoorporativaAPI.Infraestructure;
 using WebCoorporativaAPI.Models;
 
@@ -22,6 +24,9 @@ namespace WebCoorporativaAPI.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            //if (!User.TienePermiso("2.consultar"))
+            //    return Forbid();
+
             var perfiles = await _perfilService.GetAll();
             return Ok(perfiles);
         }
@@ -29,6 +34,9 @@ namespace WebCoorporativaAPI.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(int id)
         {
+            //if (!User.TienePermiso("2.consultar"))
+            //    return Forbid();
+
             var perfil = await _perfilService.GetById(id);
             if (perfil == null)
             {
@@ -43,6 +51,10 @@ namespace WebCoorporativaAPI.Controllers
         [HttpPost]
         public async Task<IActionResult> Create(PerfilModel perfil)
         {
+
+            //if (!User.TienePermiso("2.agregar"))
+              //  return Forbid();
+
             var result = await _perfilService.Create(perfil);
             if (result == null)
             {
@@ -57,6 +69,9 @@ namespace WebCoorporativaAPI.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(int id, PerfilModel perfil)
         {
+            if (!User.TienePermiso("2.editar"))
+                return Forbid();
+
             var result = await _perfilService.Update(id, perfil);
             if (!result)
             {
@@ -71,6 +86,9 @@ namespace WebCoorporativaAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!User.TienePermiso("2.eliminar"))
+                return Forbid();
+
             var result = await _perfilService.Delete(id);
             if (!result)
             {
